// src/types.ts
export interface Psychologist {
  id: number;
  nume: string;
  specialitate: string;
  treapta_specializare: string;
  regim_exercitare: string;
  status: string;
  cod_personal: string;
  numar_atestat: number;
  data_eliberare_atestat: string;
  filiala: string;
  comisia_de_avizare: string;
  dgpc: boolean;
  tsa: boolean;
  expert: boolean;
  email: string;
}